+++ library/src/actions/types.ts
@@ -1,7 +1,7 @@
 /**
  * Array requirement type.
  */
-export type ArrayRequirement<TInput extends unknown[]> = (
+export type ArrayRequirement<TInput extends readonly unknown[]> = (
   item: TInput[number],
   index: number,
   array: TInput
@@ -16,7 +16,7 @@ export type ContentInput = string | unknown[];
  * Content requirement type.
  */
 export type ContentRequirement<TInput extends ContentInput> =
-  TInput extends unknown[] ? TInput[number] : TInput;
+  TInput extends readonly unknown[] ? TInput[number] : TInput;
 
 /**
  * Length input type.