+++ library/src/methods/required/requiredAsync.ts
@@ -23,7 +23,6 @@ import type {
   InferObjectInput,
   InferObjectOutput,
   InferOutput,
-  MaybeReadonly,
   NoPipe,
   ObjectEntriesAsync,
   ObjectKeys,
@@ -55,10 +54,10 @@ type Schema = NoPipe<
  */
 type RequiredEntries<
   TEntries extends ObjectEntriesAsync,
-  TKeys extends MaybeReadonly<(keyof TEntries)[]> | undefined,
+  TKeys extends readonly (keyof TEntries)[] | undefined,
   TMessage extends ErrorMessage<NonOptionalIssue> | undefined,
 > = {
-  [TKey in keyof TEntries]: TKeys extends MaybeReadonly<(keyof TEntries)[]>
+  [TKey in keyof TEntries]: TKeys extends readonly (keyof TEntries)[]
     ? TKey extends TKeys[number]
       ? NonOptionalSchemaAsync<TEntries[TKey], TMessage>
       : TEntries[TKey]