+++ website/src/routes/api/(actions)/someItem/properties.ts
@@ -5,6 +5,7 @@ export const properties: Record<string, PropertyProps> = {
     modifier: 'extends',
     type: {
       type: 'array',
+      modifier: 'readonly',
       item: 'unknown',
     },
   },