+++ website/src/routes/api/(types)/ArrayPathItem/properties.ts
@@ -15,8 +15,15 @@ export const properties: Record<string, PropertyProps> = {
   },
   input: {
     type: {
-      type: 'array',
-      item: 'unknown',
+      type: 'custom',
+      name: 'MaybeReadonly',
+      href: '../MaybeReadonly/',
+      generics: [
+        {
+          type: 'array',
+          item: 'unknown',
+        },
+      ],
     },
   },
 };